Move odometry for classic and omni into separate files to avoid parameter override
